separate prox module
